+++ terraform/modules/happy-env-eks/namespace.tf
@@ -1,5 +1,9 @@
+locals {
+  k8s_namespace = "${var.tags.project}-${var.tags.env}-${var.tags.service}-happy-env"
+}
+
 resource "kubernetes_namespace" "happy" {
   metadata {
-    name = "${var.tags.project}-${var.tags.env}-${var.tags.service}-happy-env"
+    name = local.k8s_namespace
   }
 }