+++ terraform/modules/happy-stack-eks/versions.tf
@@ -12,7 +12,6 @@ terraform {
       source  = "datadog/datadog"
       version = ">= 3.20.0"
     }
-
     validation = {
       source  = "tlkamp/validation"
       version = "1.0.0"