+++ terraform/modules/happy-stack-eks/CHANGELOG.md
@@ -1,5 +1,52 @@
 # Changelog
 
+## [2.4.0](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v2.3.0...happy-stack-eks-v2.4.0) (2023-02-03)
+
+
+### Features
+
+* Sample Happy Environment EKS Datadog dashboard ([#1066](https://github.com/chanzuckerberg/happy/issues/1066)) ([b4c9f3f](https://github.com/chanzuckerberg/happy/commit/b4c9f3fb7df7d131093a282cb2b54fe83f1e5143))
+
+## [2.3.0](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v2.2.0...happy-stack-eks-v2.3.0) (2023-02-02)
+
+
+### Features
+
+* Add grouping tags ([#1060](https://github.com/chanzuckerberg/happy/issues/1060)) ([713015f](https://github.com/chanzuckerberg/happy/commit/713015ff7c24278c6315b9ad0ce04e98fb56bb4e))
+* Create stack level dashboard ([#1062](https://github.com/chanzuckerberg/happy/issues/1062)) ([b346da9](https://github.com/chanzuckerberg/happy/commit/b346da951e14f4c10ab7c9a936990d47913b4c92))
+
+## [2.2.0](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v2.1.1...happy-stack-eks-v2.2.0) (2023-02-01)
+
+
+### Features
+
+* Host level routing assigns an incorrect group name to an ingress, causing ALB not to be provisioned ([#1054](https://github.com/chanzuckerberg/happy/issues/1054)) ([16432eb](https://github.com/chanzuckerberg/happy/commit/16432eb0f5170f4ae601f460e8e73abdc0a747b0))
+
+## [2.1.1](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v2.1.0...happy-stack-eks-v2.1.1) (2023-02-01)
+
+
+### Bug Fixes
+
+* Force-release happy-stack-eks ([#1052](https://github.com/chanzuckerberg/happy/issues/1052)) ([6e41fe9](https://github.com/chanzuckerberg/happy/commit/6e41fe95a43aa19e3867127d5a4596b3ca62c2ab))
+
+## [2.1.0](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v2.0.0...happy-stack-eks-v2.1.0) (2023-02-01)
+
+
+### Features
+
+* Enable mapping of additional environment variables from secrets ([#1046](https://github.com/chanzuckerberg/happy/issues/1046)) ([6ef2fea](https://github.com/chanzuckerberg/happy/commit/6ef2feaf13d07a7848f8498ed14653610d1edc94))
+
+## [2.0.0](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v1.8.0...happy-stack-eks-v2.0.0) (2023-01-31)
+
+
+### ⚠ BREAKING CHANGES
+
+* authenticate ALBs for ingresses ([#1021](https://github.com/chanzuckerberg/happy/issues/1021))
+
+### Features
+
+* authenticate ALBs for ingresses ([#1021](https://github.com/chanzuckerberg/happy/issues/1021)) ([7cd9375](https://github.com/chanzuckerberg/happy/commit/7cd937576a11b16cbf07e3babf268649c48c0976))
+
 ## [1.8.0](https://github.com/chanzuckerberg/happy/compare/happy-stack-eks-v1.7.0...happy-stack-eks-v1.8.0) (2023-01-25)